# public/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Storage API",
        "description": "This is documentation for API",
        "contact": {
            "email": "apiteam@wordnik.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "depot.esmc.info",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/depot_organization_roles/": {
            "get": {
                "tags": [
                    "Depot organization role"
                ],
                "summary": "Return all depot organization roles with pagination",
                "description": "This is method for find all depot organization roles",
                "operationId": "findAllDepotOrganizationRoles",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/DepotOrganizationRole"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Depot organization role"
                ],
                "summary": "Create new depot organization role",
                "description": "This is method create new depot organization role",
                "operationId": "storeDepotOrganizationRole",
                "parameters": [
                    {
                        "name": "depot_id",
                        "in": "query",
                        "description": "ID of depot",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "organization_id",
                        "in": "query",
                        "description": "ID of organization",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "user_role_id",
                        "in": "query",
                        "description": "ID of user role",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Role",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/DepotOrganizationRole"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/depot_organization_roles/{id}": {
            "get": {
                "tags": [
                    "Depot organization role"
                ],
                "summary": "Return one depot organization role",
                "description": "This is method for find one depot organization role by ID",
                "operationId": "findDepotOrganizationRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/DepotOrganizationRole"
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Depot organization role"
                ],
                "summary": "Update depot organization role",
                "description": "This is method update depot organization role by ID",
                "operationId": "updateDepotOrganizationRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of depot organization role",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "depot_id",
                        "in": "query",
                        "description": "ID of depot",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "organization_id",
                        "in": "query",
                        "description": "ID of organization",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "user_role_id",
                        "in": "query",
                        "description": "ID of user role",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Role",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/DepotOrganizationRole"
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Depot organization role"
                ],
                "summary": "Remove depot organization role by ID",
                "description": "This is method remove depot organization role",
                "operationId": "destroyDepotOrganizationRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID depot organization role for remove",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "bad request"
                    }
                }
            }
        },
...
        "/itemgroups/{id}": {
            "get": {
                "tags": [
                    "Item group"
                ],
                "summary": "Return one item group",
                "description": "This is method for find one item group by ID",
                "operationId": "findItemGroup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/ItemGroup"
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Depot organization role"
                ],
                "summary": "Update depot organization role",
                "description": "This is method update depot organization role by ID",
                "operationId": "updateDepotOrganizationRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of item group",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of item group",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "alias",
                        "in": "query",
                        "description": "Alias of item group",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "is_available",
                        "in": "query",
                        "description": "ID of user role",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Parent ID item group",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "item_attributes",
                        "in": "query",
                        "description": "Array IDs item attributes",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "schema": {
                            "title": "data",
                            "type": "object",
                            "$ref": "#/definitions/ItemGroup"
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Item group"
                ],
                "summary": "Remove item group by ID",
                "description": "This is method remove item group",
                "operationId": "destroyItemGroup",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID item group for remove",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "bad request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "DepotOrganizationRole": {
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "depot_id": {
                                "type": "integer"
                            },
                            "organization_id": {
                                "type": "integer"
                            },
                            "user_role_id": {
                                "type": "integer"
                            },
                            "role": {
                                "type": "string"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "dateTime"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "dateTime"
                            }
                        }
                    }
                }
            },
            "type": "object"
        },
...
        "ItemGroup": {
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "alias": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            },
                            "description": {
                                "type": "string"
                            },
                            "is_available": {
                                "type": "boolean"
                            },
                            "parent_id": {
                                "type": "integer"
                            },
                            "depth": {
                                "type": "integer"
                            }
                        }
                    }
                }
            },
            "type": "object"
        },
...
    },
...
}